Изменение цели проекта с обычного бота на бота гильдии "Protego Totalum" (#4)
