+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,14 +1,15 @@
 ---
-name: 🪲 Bugs report
-about: Создайте отчет об ошибке, чтобы мы могли заняться улучшением
-title: '[BUG] Короткое название ошибки.'
-labels: 'type: bug, priority: low, status: backlog'
-assignees: 'SmithyVL'
-
+name: 🪲 Ошибка
+about: Создайте задачу с подробным описанием ошибки, чтобы мы могли заняться исправлением
+title: "Короткое название"
+labels: ["type: bug"]
+projects: ["HogwartsSchoolOfMagic/projects/2"]
+assignees: 
+  - SmithyVL
 ---
 
 **Описание ошибки**
-Четкое и краткое описание ошибки.
+Четкое и краткое описание.
 
 **Воспроизведение**
 Шаги по воспроизведению поведения:
@@ -27,4 +28,4 @@ assignees: 'SmithyVL'
 Если возможно, пожалуйста, прикрепите скриншоты, чтобы объяснить вашу проблему.
 
 **Дополнительный контекст**
-Добавьте сюда любой другой контекст проблемы.
+Добавьте сюда любой другой контекст.
