+++ .github/ISSUE_TEMPLATE/configuration_issues.md
@@ -1,14 +1,15 @@
 ---
-name: 🧰 Configuration issues
-about: Различные предложения по настройке проекта
-title: '[CONFIGURATION] Краткое название новой конфигурации.'
-labels: 'type: configuration, status: backlog, priority: low'
-assignees: 'SmithyVL'
-
+name: 🧰 Конфигурация
+about: Различные предложения по настройке проекта и сервисов внутри него
+title: "Краткое название"
+labels: ["type: configuration"]
+projects: ["HogwartsSchoolOfMagic/projects/2"]
+assignees: 
+  - SmithyVL
 ---
 
 **Описание конфигурации**
-Четкое и краткое описание того, что вы хотите, чтобы произошло.
+Четкое и краткое описание того, что вы хотите, чтобы добавилось.
 
 **Опишите альтернативные варианты**
 Четкое и краткое описание любых альтернативных решений или функций, которые вы рассмотрели.
