+++ .github/workflows/ci.yml
@@ -2,9 +2,6 @@ name: CI
 
 on:
   workflow_dispatch:
-  #push:
-  #  branches:
-  #    - master
   pull_request:
     types: [ opened, synchronize, reopened ]
 
@@ -14,11 +11,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: 📦 Checkout
-        uses: actions/checkout@v3
-      - name: ☕️ Установка JDK 17
-        uses: actions/setup-java@v3
+        uses: actions/checkout@v4
+      - name: ☕️ Установка JDK 21
+        uses: actions/setup-java@v4
         with:
-          java-version: 17
+          java-version: 21
           distribution: "temurin"
           cache: 'gradle'
       - name: 🔓 Установка разрешений на gradlew
@@ -31,11 +28,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: 📦 Checkout
-        uses: actions/checkout@v3
-      - name: ☕️ Установка JDK 17
-        uses: actions/setup-java@v3
+        uses: actions/checkout@v4
+      - name: ☕️ Установка JDK 21
+        uses: actions/setup-java@v4
         with:
-          java-version: 17
+          java-version: 21
           distribution: "temurin"
           cache: 'gradle'
       - name: 🔓 Установка разрешений на gradlew
@@ -48,19 +45,19 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: 📦 Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
         with:
           fetch-depth: 0 # Shallow clones should be disabled for a better relevancy of analysis
-      - name: ☕️ Установка JDK 17
-        uses: actions/setup-java@v3
+      - name: ☕️ Установка JDK 21
+        uses: actions/setup-java@v4
         with:
-          java-version: 17
+          java-version: 21
           distribution: "temurin"
           cache: 'gradle'
       - name: 🔓 Установка разрешений на gradlew
         run: chmod +x ./gradlew
       - name: 📒 Кэширование SonarCloud
-        uses: actions/cache@v3
+        uses: actions/cache@v4
         with:
           path: ~/.sonar/cache
           key: ${{ runner.os }}-sonar
