+++ .idea/codeStyles/Project.xml
@@ -1,0 +1,15 @@
+<component name="ProjectCodeStyleConfiguration">
+    <code_scheme name="Project" version="173">
+        <JetCodeStyleSettings>
+            <option name="PACKAGES_TO_USE_STAR_IMPORTS">
+                <value />
+            </option>
+            <option name="NAME_COUNT_TO_USE_STAR_IMPORT" value="2147483647" />
+            <option name="NAME_COUNT_TO_USE_STAR_IMPORT_FOR_MEMBERS" value="2147483647" />
+            <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
+        </JetCodeStyleSettings>
+        <codeStyleSettings language="kotlin">
+            <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
+        </codeStyleSettings>
+    </code_scheme>
+</component>
