+++ .idea/codeStyles/codeStyleConfig.xml
@@ -1,0 +1,5 @@
+<component name="ProjectCodeStyleConfiguration">
+    <state>
+        <option name="USE_PER_PROJECT_SETTINGS" value="true" />
+    </state>
+</component>
