--- README.md
@@ -1,42 +1,0 @@
-<!--suppress HtmlDeprecatedAttribute -->
-<div align="center">
-    <h1>
-        <a href="https://hogwartsschoolofmagic.github.io/DiscordLeoBot/">LeoBot для Discord</a>
-    </h1>
-</div>
-
-<div align="center">
-    <img src="https://img.shields.io/github/last-commit/HogwartsSchoolOfMagic/DiscordLeoBot?style=for-the-badge" height="25" alt="last-commit" />
-    <a href="https://wakatime.com/@SmithyVL"><img src="https://wakatime.com/badge/github/HogwartsSchoolOfMagic/DiscordLeoBot.svg?style=for-the-badge" height="25" alt="time-with-code" /></a>
-    <a href="https://github.com/HogwartsSchoolOfMagic/DiscordLeoBot/actions/workflows/ci.yml"><img src="https://img.shields.io/github/actions/workflow/status/HogwartsSchoolOfMagic/DiscordLeoBot/ci.yml?branch=master&label=CI&style=for-the-badge" height="25" alt="ci" /></a>
-    <a href="https://sonarcloud.io/code?id=HogwartsSchoolOfMagic_DiscordLeoBot"><img src="https://sonarcloud.io/api/project_badges/measure?project=HogwartsSchoolOfMagic_DiscordLeoBot&metric=ncloc" height="25" alt="sonar-code-lines" /></a>
-    <a href="https://sonarcloud.io/summary/new_code?id=HogwartsSchoolOfMagic_DiscordLeoBot"><img src="https://sonarcloud.io/api/project_badges/measure?project=HogwartsSchoolOfMagic_DiscordLeoBot&metric=alert_status" height="25" alt="sonar-quality-gate-status" /></a>
-</div>
-
-### 📖 Описание
-
-___
-Пока не написано
-
-### 📋 Инструкции по запуску
-
-___
-Пока не написано
-
-### ⚙ Включает функциональность
-
-___
-Пока не написано
-
-### 🔨 Стек технологий
-
-___
-Пока не написано
-
-### 🎫 Лицензия
-
-___
-
-**[Apache License Version 2.0](https://github.com/HogwartsSchoolOfMagic/ConfigurationService/blob/master/LICENSE)**
-
-_Copyright ©2023, Владислав [[SmithyVL]](https://github.com/SmithyVL) Кузнецов_
