+++ docs/README.md
@@ -1,0 +1,43 @@
+<!--suppress HtmlDeprecatedAttribute -->
+<div align="center">
+    <h1>
+        <a href="https://hogwartsschoolofmagic.github.io/ProtegoTotalumBot/">
+            Discord бот гильдии "Protego Totalum"
+        </a>
+    </h1>
+</div>
+
+<div align="center">
+    <img src="https://img.shields.io/github/last-commit/HogwartsSchoolOfMagic/ProtegoTotalumBot?style=for-the-badge" height="25" alt="last-commit" />
+    <a href="https://github.com/HogwartsSchoolOfMagic/ProtegoTotalumBot/actions/workflows/ci.yml"><img src="https://img.shields.io/github/actions/workflow/status/HogwartsSchoolOfMagic/ProtegoTotalumBot/ci.yml?branch=master&label=CI&style=for-the-badge" height="25" alt="ci" /></a>
+    <a href="https://sonarcloud.io/code?id=HogwartsSchoolOfMagic_ProtegoTotalumBot"><img src="https://sonarcloud.io/api/project_badges/measure?project=HogwartsSchoolOfMagic_ProtegoTotalumBot&metric=ncloc" height="25" alt="sonar-code-lines" /></a>
+    <a href="https://sonarcloud.io/summary/new_code?id=HogwartsSchoolOfMagic_ProtegoTotalumBot"><img src="https://sonarcloud.io/api/project_badges/measure?project=HogwartsSchoolOfMagic_ProtegoTotalumBot&metric=alert_status" height="25" alt="sonar-quality-gate-status" /></a>
+</div>
+
+### 📖 Описание
+
+___
+Пока не написано
+
+### 📋 Инструкции по запуску
+
+___
+Пока не написано
+
+### ⚙ Включает функциональность
+
+___
+Пока не написано
+
+### 🔨 Стек технологий
+
+___
+Пока не написано
+
+### 🎫 Лицензия
+
+___
+
+**[Apache License Version 2.0](https://github.com/HogwartsSchoolOfMagic/ConfigurationService/blob/master/LICENSE)**
+
+_Copyright ©2023, Владислав [[SmithyVL]](https://github.com/SmithyVL) Кузнецов_
