+++ pt-database/src/main/kotlin/io/github/hogwartsschoolofmagic/config/DatabaseAutoConfiguration.kt
