+++ pt-discord/src/main/kotlin/io/github/hogwartsschoolofmagic/config/DiscordConfig.kt
