+++ pt-discord/src/main/resources/application.yml
